[fix] booking/bff 패키지 -> booking 패키지로 통합 #182
