+++ booking/src/main/java/com/pyokemon/booking/dto/response/TotalSoldTicketsResponseDto.java
@@ -1,4 +1,4 @@
-package com.pyokemon.booking.bff.dto;
+package com.pyokemon.booking.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
