+++ booking/src/main/resources/mapper/BookingMapper.xml
@@ -33,7 +33,7 @@
         LIMIT 1
     </select>
 
-    <select id="findById" resultType="Booking">
+    <select id="findByBookingId" resultType="Booking">
         SELECT booking_id, event_schedule_id, seat_id, account_id, tenant_id, payment_id, status, created_at, updated_at
         FROM tb_booking
         WHERE booking_id = #{bookingId}
@@ -87,6 +87,100 @@
         FROM tb_booking
         WHERE event_schedule_id = #{eventScheduleId}
         AND status = 'BOOKED'
+    </select>
+
+    <select id="findByEventScheduleId" resultType="Booking">
+        SELECT booking_id, event_schedule_id, seat_id, account_id, tenant_id, payment_id, status, created_at, updated_at
+        FROM tb_booking
+        WHERE event_schedule_id = #{eventScheduleId}
+        ORDER BY created_at DESC
+    </select>
+
+    <select id="findByAccountIdOrderByDate" resultType="Booking">
+        SELECT booking_id, event_schedule_id, seat_id, account_id, tenant_id, payment_id, status, created_at, updated_at
+        FROM tb_booking
+        WHERE account_id = #{accountId}
+          AND status IN ('BOOKED', 'CANCELED')
+        ORDER BY created_at DESC
+            LIMIT #{size} OFFSET #{offset}
+    </select>
+
+    <select id="findByEventScheduleIdOrderByBookingId" resultType="Booking">
+        SELECT booking_id, event_schedule_id, seat_id, account_id, tenant_id, payment_id, status, created_at, updated_at
+        FROM tb_booking
+        WHERE event_schedule_id = #{eventScheduleId}
+        ORDER BY booking_id DESC
+            LIMIT #{size} OFFSET #{offset}
+    </select>
+
+    <select id="countByAccountId" resultType="long">
+        SELECT COUNT(*)
+        FROM tb_booking
+        WHERE account_id = #{accountId}
+    </select>
+
+    <select id="countByEventScheduleId" resultType="long">
+        SELECT COUNT(*)
+        FROM tb_booking
+        WHERE event_schedule_id = #{eventScheduleId}
+    </select>
+
+    <select id="findBookingCountsByScheduleIds" resultType="com.pyokemon.booking.dto.response.BookingCountDto">
+        SELECT
+        event_schedule_id AS eventScheduleId,
+        COUNT(*) AS ticketCount
+        FROM
+        tb_booking
+        WHERE
+        status = 'BOOKED'
+        AND event_schedule_id IN
+        <foreach collection="scheduleIds" item="id" open="(" separator="," close=")">
+            #{id}
+        </foreach>
+        GROUP BY
+        event_schedule_id
+    </select>
+
+    <select id="countTotalSoldTicketsByScheduleIds" resultType="long">
+        SELECT
+        COUNT(*)
+        FROM
+        tb_booking
+        WHERE
+        status = 'BOOKED'
+        AND event_schedule_id IN
+        <foreach collection="scheduleIds" item="id" open="(" separator="," close=")">
+            #{id}
+        </foreach>
+    </select>
+
+    <select id="findByAccountWithCursor" resultType="Booking">
+        SELECT booking_id, event_schedule_id, tenant_id
+        FROM tb_booking
+        WHERE account_id = #{accountId}
+        <if test="cursor != null">
+            AND booking_id &lt; #{cursor}
+        </if>
+        ORDER BY booking_id DESC
+        LIMIT #{size}
+    </select>
+
+    <!-- 2) 특정 스케줄 집합 + 계정 예매순 커서 -->
+    <select id="findByAccountAndSchedulesWithCursor" resultType="Booking">
+        SELECT booking_id, event_schedule_id, tenant_id
+        FROM tb_booking
+        WHERE account_id = #{accountId}
+        <if test="scheduleIds != null and scheduleIds.size() &gt; 0">
+            AND event_schedule_id IN
+            <foreach collection="scheduleIds" item="id" open="(" separator="," close=")">
+                #{id}
+            </foreach>
+        </if>
+        <if test="cursor != null">
+            AND booking_id &lt; #{cursor}
+        </if>
+        ORDER BY booking_id DESC
+        LIMIT #{size}
     </select>
 
 </mapper>
